:recycle: refactor: 클래스 네이밍 변경 및 도메인 구조 변경 (#13)
