+++ src/main/java/com/adit/backend/infra/crawler/platform/BrunchCrawlingStrategy.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Component;
 
 import com.adit.backend.domain.ai.dto.response.CrawlCompletionResponse;
 import com.adit.backend.global.error.GlobalErrorCode;
+import com.adit.backend.infra.crawler.WebContentCrawler;
 import com.adit.backend.infra.crawler.common.AbstractWebCrawlingStrategy;
 import com.adit.backend.infra.crawler.exception.CrawlingException;
-import com.adit.backend.infra.crawler.util.CrawlingUtil;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -45,16 +45,16 @@ public class BrunchCrawlingStrategy extends AbstractWebCrawlingStrategy {
 		StringBuilder contentBuilder = new StringBuilder();
 
 		try {
-			CrawlingUtil.extractTitle(document, TITLE_TAG, contentBuilder);
+			WebContentCrawler.extractTitle(document, TITLE_TAG, contentBuilder);
 			Elements contentElements = selectContentElements(document);
 			if (!contentElements.isEmpty()) {
 				Element mainContent = contentElements.first();
-				CrawlingUtil.extractBodyText(mainContent, TEXT_TAG, MINIMUM_RECOGNIZED_CHARACTER, contentBuilder);
+				WebContentCrawler.extractBodyText(mainContent, TEXT_TAG, MINIMUM_RECOGNIZED_CHARACTER, contentBuilder);
 			}
-			String content = CrawlingUtil.preprocessText(contentBuilder.toString());
-			String placeInfo = CrawlingUtil.extractPlaceInfo(document);
+			String content = WebContentCrawler.preprocessText(contentBuilder.toString());
+			String placeInfo = WebContentCrawler.extractPlaceInfo(document);
 			String combined = content + PLACE_SEPARATOR + placeInfo;
-			return CrawlingUtil.getCrawlCompletionResponse(contentElements, combined);
+			return WebContentCrawler.getCrawlCompletionResponse(contentElements, combined);
 		} catch (Exception e) {
 			log.error("[본문 추출 중 오류 발생] : {}", e.getMessage());
 			throw new CrawlingException(GlobalErrorCode.CRAWLING_FAILED);
